Update Requirements.txt
Here is what the corrected requirements.txt tells the installer to do:

streamlit

pandas

scikit-learn

textblob

matplotlib

spacy>=3.7.0,<4.0.0

Instruction: "Go to the official PyPI repository and install these standard libraries. For spacy, make sure to get a version that is 3.7.0 or newer, but not as new as version 4.0.0."

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

Instruction: "This is not on PyPI. Go to this specific URL on GitHub and download the file you find there. This file is the English language model (en_core_web_sm) that SpaCy needs to function."